requirements.txt
# TensorTrade Core
numpy>=1.26.4
pandas>=2.2.3,<3.0
gymnasium>=0.28.1
pyyaml>=5.1.2
stochastic>=0.6.0
tensorflow>=2.15.1
ipython>=7.12.0
matplotlib>=3.1.1
plotly>=4.5.0
ipywidgets>=7.0.0
deprecated>=1.2.13
ta>=0.4.7
pytest>=7.0.0

# Django Full-Stack Web Application
django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
# psycopg2-binary>=2.9.9  # Uncomment for PostgreSQL

# Real-time features (optional)
# redis>=5.0.0
# celery>=5.3.0
# channels>=4.0.0

# AI/ML Dependencies
# torch>=2.0.0  # Uncomment if using PyTorch
# stable-baselines3>=2.0.0  # Uncomment for RL agents
# ray[rllib]>=2.0.0  # Uncomment for distributed training
